+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 2.1.13
 ++++++
 * Add support for VNet/IP rules.
+* Add --platform parameter for `acr build`, `acr task create` and `acr task update` commands
 
 2.1.12
 ++++++