+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,7 +2,12 @@
 
 Release History
 ===============
+0.2.12
+++++++
+* webapp: add support for updating configurations through json
+* appservice plan: Updating help for appservice-plan-update command
 * functionapp: add support for app insights on functionapp create
+* webapp: bugfixes for webapp ssh
 
 0.2.11
 ++++++