+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,6 +2,8 @@
 
 Release History
 ===============
+* vm list-skus: Allow use of `--all` in place of `--all true`
+
 2.2.13
 ++++++
 * `vm encryption enable`: now validates --disk encryption keyvault and that key encryption keyvault exists, in line with `vmss encryption enable`. Exposes --force. For more info: see issues #8111 and #8110